Fix flake8 errors in Github Action.
The poetry.toml configuration was creating a virtualenv in
the project directory and this was causing flake8 to report errors
in site-packages within the venv.
This caused builds to fail in Github actions.
So as a belt and braces approach poetry.toml is now ignored
and poetry is installed and run after the flake8 tests.